fix: variable token duplicada en register + @Modifying deleteOldTokens debe ser int
- AuthService.register: renombrar 'token' (invitationToken) → 'invitationToken'
  para evitar conflicto con la variable JWT 'token' en el mismo método
- PasswordResetTokenRepository.deleteOldTokens: cambiar long → int/Integer
  que es el único tipo permitido por Spring Data en métodos @Modifying

+++ src/main/java/com/matchmetrics/persistence/repository/PasswordResetTokenRepository.java
@@ -23,9 +23,9 @@ public interface PasswordResetTokenRepository extends JpaRepository<PasswordRese
 
     @Modifying
     @Query("DELETE FROM PasswordResetToken t WHERE t.expiresAt < :cutoff OR (t.used = true AND t.createdAt < :cutoff)")
-    long deleteOldTokens(@Param("cutoff") LocalDateTime cutoff);
+    int deleteOldTokens(@Param("cutoff") LocalDateTime cutoff);
 
-    default long deleteOldTokens() {
+    default int deleteOldTokens() {
         return deleteOldTokens(LocalDateTime.now().minusHours(24));
     }
 }
